content/canvas/test/test_2d.imageData.object.ctor.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.object.ctor</title>
<!-- Testing: ImageData does not have a usable constructor -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

todo(window.ImageData !== undefined, "window.ImageData !== undefined");
try { var _thrown = false;
  new window.ImageData(1,1);
} catch (e) { _thrown = true; } finally { ok(_thrown, "should throw exception"); }

SimpleTest.finish();

});
</script>
